+++ app/client/src/sagas/InitSagas.ts
@@ -19,6 +19,8 @@ import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
 import { ERROR_CODES } from "@appsmith/constants/ApiConstants";
 import { resetApplicationWidgets, resetPageList } from "actions/pageActions";
 import { resetCurrentApplication } from "@appsmith/actions/applicationActions";
+import log from "loglevel";
+import * as Sentry from "@sentry/react";
 import { resetRecentEntities } from "actions/globalSearchActions";
 import { resetEditorSuccess } from "actions/initActions";
 import {
@@ -29,8 +31,13 @@ import {
 import { getIsInitialized as getIsViewerInitialized } from "selectors/appViewSelectors";
 import { enableGuidedTour } from "actions/onboardingActions";
 import { setPreviewModeAction } from "actions/editorActions";
+import type { AppEnginePayload } from "entities/Engine";
+import type AppEngine from "entities/Engine";
+import { AppEngineApiError } from "entities/Engine";
+import AppEngineFactory from "entities/Engine/factory";
 import type { ApplicationPagePayload } from "@appsmith/api/ApplicationApi";
 import { updateSlugNamesInURL } from "utils/helpers";
+import { generateAutoHeightLayoutTreeAction } from "actions/autoHeightActions";
 
 export const URL_CHANGE_ACTIONS = [
   ReduxActionTypes.CURRENT_APPLICATION_NAME_UPDATE,
@@ -67,6 +74,37 @@ export function* failFastApiCalls(
     return false;
   }
   return true;
+}
+
+export function* startAppEngine(action: ReduxAction<AppEnginePayload>) {
+  try {
+    const engine: AppEngine = AppEngineFactory.create(
+      action.payload.mode,
+      action.payload.mode,
+    );
+    engine.startPerformanceTracking();
+    yield call(engine.setupEngine, action.payload);
+    const { applicationId, toLoadPageId } = yield call(
+      engine.loadAppData,
+      action.payload,
+    );
+    yield call(engine.loadAppURL, toLoadPageId, action.payload.pageId);
+    yield call(engine.loadAppEntities, toLoadPageId, applicationId);
+    yield call(engine.loadGit, applicationId);
+    yield call(engine.completeChore);
+    yield put(generateAutoHeightLayoutTreeAction(true, false));
+    engine.stopPerformanceTracking();
+  } catch (e) {
+    log.error(e);
+    if (e instanceof AppEngineApiError) return;
+    Sentry.captureException(e);
+    yield put({
+      type: ReduxActionTypes.SAFE_CRASH_APPSMITH_REQUEST,
+      payload: {
+        code: ERROR_CODES.SERVER_ERROR,
+      },
+    });
+  }
 }
 
 function* resetEditorSaga() {
@@ -122,8 +160,22 @@ function* updateURLSaga(action: ReduxURLChangeAction) {
   });
 }
 
+function* appEngineSaga(action: ReduxAction<AppEnginePayload>) {
+  yield race({
+    task: call(startAppEngine, action),
+    cancel: take(ReduxActionTypes.RESET_EDITOR_REQUEST),
+  });
+}
+
 export default function* watchInitSagas() {
   yield all([
+    takeLatest(
+      [
+        ReduxActionTypes.INITIALIZE_EDITOR,
+        ReduxActionTypes.INITIALIZE_PAGE_VIEWER,
+      ],
+      appEngineSaga,
+    ),
     takeLatest(ReduxActionTypes.RESET_EDITOR_REQUEST, resetEditorSaga),
     takeEvery(URL_CHANGE_ACTIONS, updateURLSaga),
   ]);
