+++ app/client/src/sagas/__tests__/initSagas.test.ts
@@ -3,7 +3,7 @@ import { APP_MODE } from "entities/App";
 import type AppEngine from "entities/Engine";
 import AppEngineFactory from "entities/Engine/factory";
 import { call } from "redux-saga/effects";
-import { startAppEngine } from "sagas/AppEngineSaga";
+import { startAppEngine } from "sagas/InitSagas";
 
 jest.mock("../../api/Api", () => ({
   __esModule: true,
